+++ src/main/java/com/mathworks/ci/builder/MatlabBuildTaskBuilder.java
@@ -1,7 +1,7 @@
 package com.mathworks.ci.builder;
 
 import com.atlassian.bamboo.specs.api.builders.task.Task;
-// import com.mathworks.ci.task.MatlabBuildTask;
+// import com.mathworks.ci.tasktype.MatlabBuildTaskType;
 import com.mathworks.ci.properties.MatlabBuildTaskProperties;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.NotNull;
