+++ src/test/java/com/mathworks/ci/MatlabReleaseInfoTest.java
@@ -2,7 +2,7 @@
  * Copyright 2021-2022 The MathWorks, Inc.
  */
 
-package com.mathworks.ci.task;
+package com.mathworks.ci;
 
 import org.junit.Test;
 import com.mathworks.ci.MatlabReleaseInfo;
