+++ src/test/resources/atlassian-plugin.xml
@@ -6,7 +6,7 @@
     </plugin-info>
 
     <!-- from our base plugin -->
-    <component-import key="MatlabCommandTask" interface="com.mathworks.ci.task.MatlabCommandTask"/>
+    <component-import key="MatlabCommandTaskType" interface="com.mathworks.ci.tasktype.MatlabCommandTaskType"/>
 
     <!-- from the product container -->
     <component-import key="applicationProperties" interface="com.atlassian.sal.api.ApplicationProperties" />
